Add embeddings.py
Loads the prepared dataframes, with the raw data transformed into a meaningful format. Prepare this data for input into the (Ro)BERT(a)
embedding layer. Includes a test pass on a generic standalone RoBERTa model.

The input preparing function may need to be moved, or the file renamed in future commits.