Initial commit for the DeepMet model
Add the DeepMet model class and some code to perform a forward pass, in order to check that it works. All components need checking for correctness and the model still needs to be trained, in order to make sure it is functioning properly/efficiently. Improvements probably necessary to make sure the model contains all of the components mentioned in the paper, but this commit should provide the backbone