Add maximum sequence length parameter
Set the max seq len param to 128, which matches the length set out in the paper. This allows us to increase the batch size to 16 (on my home PC), which also matches the value seen in the paper.

This raises some issues about truncation, since the natural max seq len for train is 201. These should be investigated later.