Break training data into fixed folds
This allows training to be split up into individual models, since we now have 10 fixed folds of data, as seen in _train_folds.csv files. Thus, we can use the same fold split each time we start training, which allows us to start/stop/distribute at will. This makes it easier to train a large number of models, which is required for cross validation. Note that some follow-up work is required, mainly in the form of tidying up and increasing customisability of training, but this commit adds the main functionality required for us to split training up a bit.